user can now change user type
